added option to link an external object to current one so that their snapshot number is in sync.
Note: so far, this works in "chains" only, i.e. one object is updated by another  one. To make it work in a bidirectional way, we need to make the list of linked objects an attribute and the update function must become also an attribute which then calls the update of all the other objects.